+++ addOns/selenium/src/main/java/org/zaproxy/zap/extension/selenium/CustomBrowser.java
@@ -20,6 +20,7 @@
 package org.zaproxy.zap.extension.selenium;
 
 import java.util.List;
+import java.util.Map;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -36,5 +37,13 @@ public class CustomBrowser {
     private final String driverPath;
     private final String binaryPath;
     private final List<String> arguments;
+
+    /**
+     * Preferences passed to the browser on startup
+     *
+     * @since 15.46.0
+     */
+    private final Map<String, String> preferences;
+
     private final String browserType;
 }